+++ src/plone/versioncheck/script.py
@@ -1,6 +1,6 @@
 # -*- coding: utf-8 -*-
 from argparse import ArgumentParser
-from plone.versioncheck.formatter import display
+from plone.versioncheck import formatter
 from plone.versioncheck.parser import parse
 from plone.versioncheck.pypi import check_all
 from plone.versioncheck import tracking
@@ -11,10 +11,10 @@ parser = ArgumentParser(
                 "simple and complex/cascaded buildouts."
 )
 parser.add_argument(
-    "-b",
-    "--buildout",
-    help='path to buildout.cfg or other cfg file',
-    default='buildout.cfg'
+    'buildout',
+    nargs='?',
+    default='buildout.cfg',
+    help='path to buildout.cfg or other *.cfg file'
 )
 parser.add_argument(
     "-p",
@@ -31,14 +31,26 @@ parser.add_argument(
 parser.add_argument(
     "-r",
     "--required-by",
-    help='show information about requirements (only with tracking possible)',
+    help='show information about requirements (only if tracking file is '
+         'available)',
     action="store_true"
 )
 parser.add_argument(
     "-i",
     "--ignore-tracking",
-    help='display only packages with overrides',
-    action="store_false"
+    help='ignore tracking file (if present)',
+    action="store_true"
+)
+parser.add_argument(
+    "-m",
+    "--machine",
+    help='show as machine readable output (json)',
+    action="store_true"
+)
+parser.add_argument(
+    '--no-colors',
+    help='do not show colors',
+    action="store_true"
 )
 parser.add_argument(
     '--debug-limit',
@@ -49,9 +61,22 @@ parser.add_argument(
 
 def run():
     args = parser.parse_args()
-    pkgsinfo = parse(args.buildout)
+    pkgsinfo = {}
+    pkgsinfo['pkgs'] = parse(args.buildout)
     if args.pypi:
         check_all(pkgsinfo, args.debug_limit)
     if not args.ignore_tracking:
         tracking.get(pkgsinfo, args.buildout)
-    display(pkgsinfo, newer_only=args.newer, limit=args.debug_limit)
+    if args.machine:
+        formatter.machine(
+            pkgsinfo,
+            newer_only=args.newer,
+            limit=args.debug_limit,
+        )
+    else:
+        formatter.human(
+            pkgsinfo,
+            newer_only=args.newer,
+            colored=args.no_colors,
+            limit=args.debug_limit,
+        )